Lookup join: move CacheInfo to utils
