+++ quixstreams/dataframe/joins/lookups/base.py
@@ -1,6 +1,6 @@
 import abc
 import dataclasses
-from typing import Any, Generic, Mapping, TypedDict, TypeVar
+from typing import Any, Generic, Mapping, TypeVar
 
 from quixstreams.models.types import HeadersMapping
 
@@ -60,19 +60,3 @@ class BaseField(abc.ABC):
     """
 
     pass
-
-
-class CacheInfo(TypedDict):
-    """
-    Typed dictionary containing cache statistics for the LRU cache.
-
-    :param hits: The number of cache hits.
-    :param misses: The number of cache misses.
-    :param size: The current size of the cache.
-    :param maxsize: The maximum size of the cache.
-    """
-
-    hits: int
-    misses: int
-    size: int
-    maxsize: int
